// src/app/core/modules/openapi/model/adres-body.ts
/**
 * Adres API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Person } from './person';


export interface AdresBody { 
    street: string;
    housenumber: string;
    zipcode?: string;
    city: string;
    persons?: Array<Person>;
}